+++ src/main/kotlin/internals/instructions/bitwise/xor.kt
@@ -16,8 +16,8 @@ import helpers.registerWrite
  */
 fun Bitwise.xor(operand1: SuperRegisterType, operand2: SuperRegisterType): Unit = try {
     registerWrite(
-        register = R3, value = registerRead(register = operand1).xor(
-            other = registerRead(register = operand2)
+        register = R3, value = (registerRead(register = operand1) as Long).xor(
+            other = registerRead(register = operand2) as Long
         )
     )
 } catch (_: Exception) {
