+++ src/main/kotlin/internals/instructions/strings/strcpy.kt
@@ -3,8 +3,8 @@ package internals.instructions.strings
 import data.memory.MemoryAddress
 import data.registers.enumIdenifiers.SuperRegisterType
 import errors
-import helpers.registerRead
 import helpers.readRegisterString
+import helpers.registerRead
 import helpers.writeStringSpecInMemory
 
 /**
@@ -17,7 +17,7 @@ import helpers.writeStringSpecInMemory
 
 fun Strings.strcpy(source: SuperRegisterType, destination: SuperRegisterType): Unit = try {
     val string: String = readRegisterString(register = source)
-    val destinationAddress: Long = registerRead(register = destination)
+    val destinationAddress: Long = registerRead(register = destination) as Long
     writeStringSpecInMemory(
         string = string, destinationAddress = MemoryAddress(address = destinationAddress)
     )
