+++ src/main/kotlin/internals/systemCalls/calls/writeFile.kt
@@ -21,14 +21,14 @@ import internals.systemCalls.SystemCall
  * @param buffer The starting address of the buffer in memory containing the data to write (stored in register S3).
  */
 fun SystemCall.writeFile(fd: SuperRegisterType, buffer: SuperRegisterType): Unit = try {
-    val f: VMFile = fileDescriptors.getFileDescriptor(fd = registerRead(register = fd))!!
+    val f: VMFile = fileDescriptors.getFileDescriptor(fd = registerRead(register = fd) as Long)!!
     var index: Int = 0
     var string: String = buildString {}
     while (true) {
 
         val byte: MemoryValue = internalMemory.read(
             address = MemoryAddress(
-                address = registerRead(register = buffer).plus(index)
+                address = (registerRead(register = buffer) as Long).plus(index)
             )
         )
         if (byte.value!!.equals(other = 0L)) {
